Restructuring of Project
For better maintainability improved Project Structure and integrated npm and parcel for project building.

Fragmentshaders for processing had been placed into own folder. Available shaders inside are gathered when "npm run build" is executed by generateShaders.js and are getting integrated into the available shaders and dropdown.
