+++ .parcelrc
@@ -1,0 +1,6 @@
+{
+  "extends": "@parcel/config-default",
+  "transformers": {
+    "*.txt": ["@parcel/transformer-inline-string"]
+  }
+}
