+++ LICENSE
@@ -2,10 +2,6 @@ MIT License
 
 Copyright (c) 2024 Tim Meyer
 
-adapted with code from
-
-Copyright (c) 2021 Rilshrink (https://github.com/Rilshrink/WebGLRipper)
-
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
 in the Software without restriction, including without limitation the rights
