+++ fragmentShader/FRAG_INVERT.txt
@@ -1,0 +1,7 @@
+precision mediump float;
+varying vec2 texCoords;
+uniform sampler2D texture;
+void main() {
+    vec4 color = texture2D(texture, texCoords);
+    gl_FragColor = vec4(vec3(1.0) - color.rgb, color.a);
+}
