+++ fragmentShader/FRAG_STANDARD.txt
@@ -1,0 +1,13 @@
+precision mediump float;
+varying vec2 texCoords;
+uniform sampler2D texture;
+void main() {
+    vec4 texColor = texture2D(texture, texCoords);
+    float brightness = max(max(texColor.r, texColor.g), texColor.b);
+    if (brightness >= 0.6) {
+        float redIntensity = (brightness - 0.6) / (1.0 - 0.6);
+        gl_FragColor = vec4(redIntensity, 0.0, 0.0, 1.0); // Proper red tone
+    } else {
+        gl_FragColor = texColor; // Original color
+    }
+}
