+++ index.html
@@ -223,7 +223,7 @@
         }
 
         window.onload = function() {
-            var baseurl = 'https://raw.githubusercontent.com/meyerstim/WebGLRipper/Developing/webglripper.js';
+            var baseurl = 'https://raw.githubusercontent.com/meyerstim/WebGLRipper/Project/dist/webglripper.min.js';
             setup_processor_button('webglripper.js', baseurl);
         }
